+++ src/GridCal/Gui/GridEditorWidget/vsc_graphics.py
@@ -20,11 +20,12 @@ from GridCal.Gui.GuiFunctions import BranchObjectModel
 from GridCal.Engine.Devices.vsc import VSC
 from GridCal.Engine.Devices.branch import Branch, BranchType
 from GridCal.Engine.Simulations.Topology.topology_driver import reduce_grid_brute
+from GridCal.Gui.GridEditorWidget.messages import *
 
 
 class VscEditor(QDialog):
 
-    def __init__(self, branch: Branch, Sbase=100):
+    def __init__(self, branch: VSC, Sbase=100):
         """
         Line Editor constructor
         :param branch: Branch object to update
@@ -207,11 +208,9 @@ class VscGraphicItem(QGraphicsLineItem):
             self.setToPort(toPort)
 
         # add transformer circles
-        self.symbol_type = BranchType.Line
-        self.symbol = None
-        self.c0 = None
-        self.c1 = None
-        self.c2 = None
+        h = 48
+        w = h
+        self.symbol = QGraphicsRectItem(QRectF(0, 0, w, h), parent=self)
         if self.api_object is not None:
             self.update_symbol()
 
@@ -225,132 +224,38 @@ class VscGraphicItem(QGraphicsLineItem):
         """
         Remove all symbols
         """
-        for elm in [self.symbol, self.c1, self.c2, self.c0]:
+        for elm in [self.symbol]:
             if elm is not None:
                 try:
                     self.diagramScene.removeItem(elm)
-                    # sip.delete(elm)
-                    elm = None
                 except:
                     pass
 
     def update_symbol(self):
         """
         Make the branch symbol
-        :return:
         """
 
         # remove the symbol of the branch
         self.remove_symbol()
+        self.make_vsc_symbol()
 
-        if self.api_object.branch_type == BranchType.Transformer:
-            self.make_transformer_symbol()
-            self.symbol_type = BranchType.Transformer
-
-        elif self.api_object.branch_type == BranchType.Switch:
-            self.make_switch_symbol()
-            self.symbol_type = BranchType.Switch
-
-        elif self.api_object.branch_type == BranchType.Reactance:
-            self.make_reactance_symbol()
-            self.symbol_type = BranchType.Switch
-
-        elif self.api_object.branch_type == BranchType.DCLine:
-            self.make_dc_line_symbol()
-            self.symbol_type = BranchType.DCLine
-
-        else:
-            # this is a line
-            self.symbol = None
-            self.c0 = None
-            self.c1 = None
-            self.c2 = None
-            self.symbol_type = BranchType.Line
-
-    def make_transformer_symbol(self):
+    def make_vsc_symbol(self):
         """
-        create the transformer simbol
-        :return:
+        Make the VSC symbol
         """
-        h = 80.0
-        w = h
-        d = w/2
-        self.symbol = QGraphicsRectItem(QRectF(0, 0, w, h), parent=self)
-        self.symbol.setPen(QPen(Qt.transparent))
-
-        self.c0 = QGraphicsEllipseItem(0, 0, d, d, parent=self.symbol)
-        self.c1 = QGraphicsEllipseItem(0, 0, d, d, parent=self.symbol)
-        self.c2 = QGraphicsEllipseItem(0, 0, d, d, parent=self.symbol)
-
-        self.c0.setPen(QPen(Qt.transparent, self.width, self.style))
-        self.c2.setPen(QPen(self.color, self.width, self.style))
-        self.c1.setPen(QPen(self.color, self.width, self.style))
-
-        self.c0.setBrush(QBrush(Qt.white))
-        self.c2.setBrush(QBrush(Qt.white))
-
-        self.c0.setPos(w * 0.35 - d / 2, h * 0.5 - d / 2)
-        self.c1.setPos(w * 0.35 - d / 2, h * 0.5 - d / 2)
-        self.c2.setPos(w * 0.65 - d / 2, h * 0.5 - d / 2)
-
-        self.c0.setZValue(0)
-        self.c1.setZValue(2)
-        self.c2.setZValue(1)
-
-    def make_switch_symbol(self):
-        """
-        Mathe the switch symbol
-        :return:
-        """
-        h = 40.0
+        h = 48
         w = h
         self.symbol = QGraphicsRectItem(QRectF(0, 0, w, h), parent=self)
         self.symbol.setPen(QPen(self.color, self.width, self.style))
+
+        graphic = QGraphicsRectItem(QRectF(0, 0, w, h), parent=self.symbol)
+        graphic.setBrush(QBrush(QPixmap(":/Icons/icons/vsc.svg")))
+
         if self.api_object.active:
             self.symbol.setBrush(self.color)
         else:
             self.symbol.setBrush(QBrush(Qt.white))
-
-    def make_dc_line_symbol(self):
-        """
-        Make the DC Line symbol
-        :return:
-        """
-        h = 30.0
-        w = h
-        w2 = int(w / 2)
-        self.symbol = QGraphicsRectItem(QRectF(0, 0, w, h), parent=self)
-
-        offset = 3
-        t_points = QPolygonF()
-        t_points.append(QPointF(0, offset))
-        t_points.append(QPointF(w-offset, w2))
-        t_points.append(QPointF(0, w-offset))
-        triangle = QGraphicsPolygonItem(self.symbol)
-        triangle.setPolygon(t_points)
-        triangle.setPen(QPen(Qt.white))
-        triangle.setBrush(QBrush(Qt.white))
-
-        line = QGraphicsRectItem(QRectF(h-offset, offset, offset, w-2*offset), parent=self.symbol)
-        line.setPen(QPen(Qt.white))
-        line.setBrush(QBrush(Qt.white))
-
-        self.symbol.setPen(QPen(self.color, self.width, self.style))
-        if self.api_object.active:
-            self.symbol.setBrush(self.color)
-        else:
-            self.symbol.setBrush(QBrush(Qt.white))
-
-    def make_reactance_symbol(self):
-        """
-        Make the reactance symbol
-        :return:
-        """
-        h = 40.0
-        w = 2 * h
-        self.symbol = QGraphicsRectItem(QRectF(0, 0, w, h), parent=self)
-        self.symbol.setPen(QPen(self.color, self.width, self.style))
-        self.symbol.setBrush(self.color)
 
     def setToolTipText(self, toolTip: str):
         """
@@ -362,11 +267,6 @@ class VscGraphicItem(QGraphicsLineItem):
 
         if self.symbol is not None:
             self.symbol.setToolTip(toolTip)
-
-        if self.c0 is not None:
-            self.c0.setToolTip(toolTip)
-            self.c1.setToolTip(toolTip)
-            self.c2.setToolTip(toolTip)
 
     def contextMenuEvent(self, event):
         """
@@ -455,53 +355,55 @@ class VscGraphicItem(QGraphicsLineItem):
         Remove this object in the diagram and the API
         @return:
         """
-        self.diagramScene.circuit.delete_branch(self.api_object)
-        self.diagramScene.removeItem(self)
+        ok = yes_no_question('Do you want to remove this VSC?', 'Remove VSC')
+
+        if ok:
+            self.diagramScene.circuit.delete_branch(self.api_object)
+            self.diagramScene.removeItem(self)
 
     def reduce(self):
         """
         Reduce this branch
         """
+        ok = yes_no_question('Do you want to reduce this VSC?', 'Remove VSC')
 
-        # get the index of the branch
-        br_idx = self.diagramScene.circuit.branches.index(self.api_object)
+        if ok:
+            # get the index of the branch
+            br_idx = self.diagramScene.circuit.branches.index(self.api_object)
 
-        # call the reduction routine
-        removed_branch, removed_bus, \
-            updated_bus, updated_branches = reduce_grid_brute(self.diagramScene.circuit, br_idx)
+            # call the reduction routine
+            removed_branch, removed_bus, \
+                updated_bus, updated_branches = reduce_grid_brute(self.diagramScene.circuit, br_idx)
 
-        # remove the reduced branch
-        removed_branch.graphic_obj.remove_symbol()
-        self.diagramScene.removeItem(removed_branch.graphic_obj)
+            # remove the reduced branch
+            removed_branch.graphic_obj.remove_symbol()
+            self.diagramScene.removeItem(removed_branch.graphic_obj)
 
-        # update the buses (the deleted one and the updated one)
-        if removed_bus is not None:
-            # merge the removed bus with the remaining one
-            updated_bus.graphic_obj.merge(removed_bus.graphic_obj)
+            # update the buses (the deleted one and the updated one)
+            if removed_bus is not None:
+                # merge the removed bus with the remaining one
+                updated_bus.graphic_obj.merge(removed_bus.graphic_obj)
 
-            # remove the updated bus children
-            for g in updated_bus.graphic_obj.shunt_children:
-                self.diagramScene.removeItem(g.nexus)
-                self.diagramScene.removeItem(g)
-            # re-draw the children
-            updated_bus.graphic_obj.create_children_icons()
+                # remove the updated bus children
+                for g in updated_bus.graphic_obj.shunt_children:
+                    self.diagramScene.removeItem(g.nexus)
+                    self.diagramScene.removeItem(g)
+                # re-draw the children
+                updated_bus.graphic_obj.create_children_icons()
 
-            # remove bus
-            for g in removed_bus.graphic_obj.shunt_children:
-                self.diagramScene.removeItem(g.nexus)  # remove the links between the bus and the children
-            self.diagramScene.removeItem(removed_bus.graphic_obj)  # remove the bus and all the children contained
+                # remove bus
+                for g in removed_bus.graphic_obj.shunt_children:
+                    self.diagramScene.removeItem(g.nexus)  # remove the links between the bus and the children
+                self.diagramScene.removeItem(removed_bus.graphic_obj)  # remove the bus and all the children contained
 
-            #
-            # updated_bus.graphic_obj.update()
-
-        for br in updated_branches:
-            # remove the branch from the schematic
-            self.diagramScene.removeItem(br.graphic_obj)
-            # add the branch to the schematic with the rerouting and all
-            self.diagramScene.parent_.add_line(br)
-            # update both buses
-            br.bus_from.graphic_obj.update()
-            br.bus_to.graphic_obj.update()
+            for br in updated_branches:
+                # remove the branch from the schematic
+                self.diagramScene.removeItem(br.graphic_obj)
+                # add the branch to the schematic with the rerouting and all
+                self.diagramScene.parent_.add_line(br)
+                # update both buses
+                br.bus_from.graphic_obj.update()
+                br.bus_to.graphic_obj.update()
 
     def remove_widget(self):
         """
@@ -627,38 +529,25 @@ class VscGraphicItem(QGraphicsLineItem):
             self.setZValue(-1)
 
             if self.api_object is not None:
+                # if the branch has a moveable symbol, move it
+                try:
+                    h = self.pos2.y() - self.pos1.y()
+                    b = self.pos2.x() - self.pos1.x()
+                    ang = np.arctan2(h, b)
+                    h2 = self.symbol.rect().height() / 2.0
+                    w2 = self.symbol.rect().width() / 2.0
+                    a = h2 * np.cos(ang) - w2 * np.sin(ang)
+                    b = w2 * np.sin(ang) + h2 * np.cos(ang)
 
-                # if the object branch type is different from the current displayed type, change it
-                if self.symbol_type != self.api_object.branch_type:
-                    self.update_symbol()
+                    center = (self.pos1 + self.pos2) * 0.5 - QPointF(a, b)
 
-                if self.api_object.branch_type == BranchType.Line:
-                    pass
+                    transform = QTransform()
+                    transform.translate(center.x(), center.y())
+                    transform.rotate(np.rad2deg(ang))
+                    self.symbol.setTransform(transform)
 
-                elif self.api_object.branch_type == BranchType.Branch:
-                    pass
-
-                else:
-
-                    # if the branch has a moveable symbol, move it
-                    try:
-                        h = self.pos2.y() - self.pos1.y()
-                        b = self.pos2.x() - self.pos1.x()
-                        ang = np.arctan2(h, b)
-                        h2 = self.symbol.rect().height() / 2.0
-                        w2 = self.symbol.rect().width() / 2.0
-                        a = h2 * np.cos(ang) - w2 * np.sin(ang)
-                        b = w2 * np.sin(ang) + h2 * np.cos(ang)
-
-                        center = (self.pos1 + self.pos2) * 0.5 - QPointF(a, b)
-
-                        transform = QTransform()
-                        transform.translate(center.x(), center.y())
-                        transform.rotate(np.rad2deg(ang))
-                        self.symbol.setTransform(transform)
-
-                    except Exception as ex:
-                        print(ex)
+                except Exception as ex:
+                    print(ex)
 
     def set_pen(self, pen):
         """
@@ -668,15 +557,8 @@ class VscGraphicItem(QGraphicsLineItem):
         """
         self.setPen(pen)
 
-        # Color the symbol only for switches
-        if self.api_object.branch_type == BranchType.Switch:
-            if self.symbol is not None:
-                self.symbol.setPen(pen)
-
-        elif self.api_object.branch_type == BranchType.Transformer:
-            if self.c1 is not None:
-                self.c1.setPen(pen)
-                self.c2.setPen(pen)
+        if self.symbol is not None:
+            self.symbol.setPen(pen)
 
     def edit(self):
         """
@@ -699,16 +581,3 @@ class VscGraphicItem(QGraphicsLineItem):
         dlg = VscEditor(self.api_object, Sbase)
         if dlg.exec_():
             pass
-
-    def tap_up(self):
-        """
-        Set one tap up
-        """
-        self.api_object.tap_up()
-
-    def tap_down(self):
-        """
-        Set one tap down
-        """
-        self.api_object.tap_down()
-
